Pass boot device as `root=` unless `root=` already given.
Telling the booted kernel which device is the root device means the
booted kernel need not figure that out by itself, possibly making ambiguous
decisions.

In most cases, this is what the user wants.

However this breaks the case where the loaded kernel and initrd are on e.g.
a `/boot` partition separate from the root partition, and the initrd has
encoded _inside it_ which device to use as root partition to use.
This is acceptable however, because such initrds have the choice to just ignore
the `root=` parameter we pass.